refactor: simplified Headers
